Revert "2 navegacion"
